Fix generate_vivado.tcl for new project, delete leftover IP products
generate_vivado.tcl under 2020.2 requires that you specify the
absolute rather than relative paths. Accordingly, people will
need to change this to get the project up and running, but the
relevant changes occur right at the beginning of the script
so no biggie.

Nits: the gitignore didn't fully ignore everything now that we
have discrete subdirectories per IP. Fix this with a big gitignore
refactor that individually ignored each subdirectory. There is
probably a better way to do this.

Signed-off-by: Jay Lang <jaytlang@mit.edu>